ajout apo pour les messages + ajout d'un id pour ceux si !!

+++ Application/UseCases/Comment/Dtos/InputDtoCreateComment.cs
@@ -3,9 +3,9 @@
     public class InputDtoCreateComment
     {
         public string Content { get; set; }
-        public int Score { get; set; }
+        public byte Score { get; set; }
         public string Image { get; set; }
-        public int Difficulty { get; set; }
+        public byte Difficulty { get; set; }
         public int IdUser { get; set; }
         public int IdRide { get; set; }
     }
