+++ Application/UseCases/Comment/Dtos/OutputDtoComment.cs
@@ -4,9 +4,9 @@
     {
         public int Id { get; set; }
         public string Content { get; set; }
-        public int Score { get; set; }
+        public byte Score { get; set; }
         public string Image { get; set; }
-        public int Difficulty { get; set; }
+        public byte Difficulty { get; set; }
         public int IdUser { get; set; }
         public int IdRide { get; set; }
     }
