+++ Domain/Comment.cs
@@ -6,11 +6,11 @@
 
         public string Content { get; set; }
 
-        public int Score { get; set; }
+        public byte Score { get; set; }
         
         public string Image { get; set; }
 
-        public int Difficulty { get; set; }
+        public byte Difficulty { get; set; }
 
         public int IdUser { get; set; }
